Replace the declarative extractor definitions with JavaScript
The approach worked technically, but turned out fairly hard to work with.
JavaScript is a bit more verbose, but easier to work with as it doesn't
enforce a very specific way of modeling the extractors. Being able to do
printf debugging inside the extractor code rather than reviewing a full
extractor rule execution trace to find mistakes is also convenient.