Introduce Flight::departureDay
This contains the scheduled departure date, as used for unique flight
identification in IATA BCBP. The actual departure time may differ from
that in case of delays beyond midnight for example.

Having this allows us to extend the IATA BCBP parser to also extract the
departure date, assuming we know when the boarding pass has been issued
(the BCBP data only contains the departure day, not the year...).