Inherit Airline from Organization
This required changing the assumption in the current data types code that
base classes are abstract. This in turn now allows assigning instances
of sub-types to instances of base types (but there's no way back, with
the polymorphism hidden in the private class). That works without the
risk of slicing as the private type will "remember" the sub-type it came
from, and can actually be quite convenient.