Start to add content-aware element merging
JsonLdDocument::apply does not take the content into account, which in
some cases leads to sub-optimal result. Instead, prefer the side that
has the "better" property value, whenever we can decide this.

So far this is only looking at QDateTime and Airport, more types to be
added.

With this the new structured data extractor using libxml2 outperforms
the old code in all test cases, including the weird Eurowings boarding
passes containing both JSON-LD and Microdata, both with different content.