Add a generic PDF extractor
This is a slightly more advanced version of what's currently in generic.js,
and scans PDF documents for supported barcodes (so far only IATA BCBP is
handled properly yet, UIC 918.3 still to be done).

The main difference however is that this is now applied on any PDF we get
our hands on, not just those whitelisted for generic.js. It will also
enable the app to import PDF files without a corresponding email context.