Add support for the newer ZXing from https://github.com/nu-book/zxing-cpp
This significantly outperforms the variant we were previously using, both
in performance and accuracy. It also doesn't have the memory leak on PDF417
detection failures observed with the other one, and has a "pure" mode
which assumes generated rather than camera input (which we have here).
This allows us to bypass the luminance source and go directly to the bit
matrix. Altogether this speeds up the test suite by 45%, while also
decoding a number of previously failing test files.