Add operator== to the schema.org types
As writing this by hand would be rather error prone and boring, this uses
a trick from Verdigris to generate them as part of the macro that generates
the getter/setter methods already, without the need of repeating the list
of properties again.

operator!= could be implemented in the same way for optimal efficiency,
but since we don't need that on a hot path right now I took the lazy
shortcut.