Add support for extracting untyped byte arrays to the extractor engine
We have that code duplicated in the unit tests, the command line extractor
and the app with various levels of precision, so better unify this here.

This will also enable the command line extractor to auto-detect the content
type on stdin (so far only implemented for PDF and PkPass), and the app to
import more content from the clipboard or content: URLs.