Determine airport country also if we have multiple IATA code candidates
As long as they are all in the same country, that is good enough. The
same will also work for timezones.

This is a common scenario with multiple airports in the same city for
example, especially when one is much more relevant than the others and
therefore often resulting in use of simplified/incomplete names.