Generalize the distance-based airport disambiguation
We can also apply this to cases where both sides are ambiguous. Such as
"Frankfurt" (resolving to FRA and HHN) to "Valencia" (resolving to VLC
and PPN). Since FRA and HHN are close to each other but VLC and PPN are
far apart we can still disambiguate VLC/PPN based on distance.