Start to make Uic9183Block safer to use
Rather than holding a raw data pointer, hold the QByteArray from the ticket
the block is part of, and an offset into that. This way it's still safe
even if the ticket goes out of scope. This also paves the way for block
iteration API needed for KItinerary Worbench.

The UIC 918.3 ticket layout code as well as the DB vendor blocks still need
to be adjusted to this.