Work towards unifying the generic extraction process
PkPass is still to be adapted to this. The general idea is to first run
whatever generic extraction we have, then select any applicable custom
extractor, and then run those with the output of the generic extraction
as context. This allows us to do custom fixups or extensions even of
structured annotations, and it should further simplify the extractor
scripts, as they no longer need to ensure their result is mergable.