Initial work on decoding VDV tickets
Those are common among German local transport providers, and visually look
like UIC 918.3 codes. Unfortunately the description of what they are valid
for is entangled with their cryptographic signatures, which makes it
necessary to deal with all that here as well to get to the information we
are interested in. I only have partial documentation for this, so this
might take a while...