Add a generic BER/DER/X.690 decoder
This will replace various incomplete adhoc solutions for this in the VDV
ticket parsing. It will also be helpful for decoding at least parts of SBB
and Thalys ticket codes.

This currently only does the bare minimum, proper type decoding and
recursing into nested structures would also be useful, as well as support
for built-in primitive type content.