Add Reservation::reservationStatus
This is needed to model cancellations, a currently unfortunately widely
needed feature. This requires a few changes to our enum (de)serialization
to be compliant with JSON-LD. And due to this having a default value it
causes quite some changes in the test files, adding the new property.

This does not actually make use of the property yet, besides from parsing
both the new and legacy forms from JSON-LD.