Initial experiments to improve airport coordinates based on OSM data
So far the coordinates we use for navigating to/from airports come from
Wikidata, and are typically somewhere on the center of the whole airport
area, typically a runway. That's far from ideal for navigation, as that's
neither a location you want to or event can get to, and in extreme cases
this even leads to navigation "snapping" to the opposite side of the
airport entirely.

Instead we want a coordinate around somewhere around the entrance. So far
there's three types of information from OSM we consider here:
* terminal buildings (or rather the center point of their bounding box)
* entrance nodes on terminal building polygons (unfortunately not reliably
available in the input data).
* railway stations on the premise of the airport.

This improves the result for many smaller or mid-sized airports
considerably already. What this cannot improve is the situation
at large airports with widely spread terminals (LHR, CDG, MXP, etc).
Those however cannot meaningfully represented by a single coordinate
anyway.

Code isn't hooked up yet, this is just for local experiments at this
point.