Change the way we store timezones in the static database
So far this were offsets into the the IANA string table, now it's a flat
enum. The old way needed 13 bits per record, the new only needs 9 bit, at
the cost of an extra ~800 bytes for an offset table to get back to IANA
names. This however quickly pays of when storing large quantities, which
we do (~37k in the current database, more in the upcoming experiments for
an efficient geo coordinate to timezone mapping).