Use both the geo coordinate and the country to determine the timezone
When running this for all train stations in our database this gets very
close to the expected timezones now, achieving 29789 matching results with
only 9 deviations (some of those are mistakes, some are in disputed areas
where the "correct" result isn't entirely obvious). Performance for
airports is slightly worse, as those are more spread out over the world,
and thus don't benefit as much from the fine-granular country structure
in Europe that helps a lot here.

There are some more ideas on how to further improve this, either way it
looks very promising as a generic replacement for the timezone information
in our static databases.