Add a IANA timezone to country mapping
This is largely unique, and should allow us to improve the result for the
coordinate-based timezone lookup when having a country context.

That was at least the initial motivation, however it looks like this might
eventually allow us to build a coordinate -> country mapping as well,
without needing an extra spatial index.