Try to find an equivalent timezone in the right country
This improves the situation for locations close to the boarder, where the
coordinate-based lookup gives us the equivalent timezone in the wrong
country, and where the right country has more than one timezone. This
particularly helps with CA <-> US and MX <-> US cases, and thus improves
test results for the airport location database.