Exclude the polar regions from the timezone index
This allows us to move about 20% of the available z-order curve length
to more relevant regions, increasing the precision there.

This however further increases the aspect ration to ~2.2, it might
therefore make sense to use an additional bit for an initial split
on the longitude to get this back to around 1. While this doesn't
generally break the approach, the spatial proximity property of the
z-order curve degrades a bit with this, reducing the encoding efficiency
of this.