Store whether we have an unambiguous or equivalent timezone
This uses the last bit we still had unused in the index, but due to the
extra information impact the run-length encoding and thus slightly
increases the number of entries.

Having this available however allows us to further increase precision,
and gives us a coordinate to country mapping essentially for free.