Switch from Gares & Connexion (P3104) to SNCF station codes (P8181)
SNCF station codes a strict superset, and also cover e.g. foreign stations
not managed by Gares & Connexions, and thus is what we actually want here
to make international tickets work as well. This only recently became
available in Wikidata though.

This does not contain the actual db update yet, that will happen as part
of the removal of the explicit timezone information.