Guard against SNCF tickets having 1st/2nd leg barcode on 3rd leg page
The barcodes can only contain two legs, and so do ticket PDF pages, so this
code was making the assumption that in a three leg case there would also be
another barcode for that page, but that doesn't seem to be the case.

In that case we then have to solely rely on the PDF page text.