Add a test case for barcodes with scaling artifacts
That is, individual lines are off by one pixel in size, making the features
in there non-square. This can be found for example in DataMatrix codes on
Renfe tickets, or QR codes on CCC tickets.

This isn't a problem for our current decoder, but fails with new ZXing
ReadBarcode() API in pure mode.