Consider horizontal flips in the PDF vector graphics transformation
We ignored this so far, possibly resulting in horizontally flipped
rastered barcodes. This went unnoticed so far as ZXing was handling
this for most codes just fine, and we handle this case explicitly for
PDF417 in our barcode decoding logic. The new ZXing API doesn't do this
anymore though, which uncovered this.

We still have the workaround for flipped PDF417 codes in place though,
as there are documents with flipped raster codes as well apparently,
but we should at least no longer hit this for vector codes now.