Make VDV ticket content structs introspectable
This allows us to print almost all content of a VDV ticket in the barcode
dumping tool, as well as the removal of a lot of diagnostic code from the
actual VDV ticket parser.

The validity area structs still need work to properly decode those, but I
yet have to find the official documents describing those.  We should have
enough information to decode 0080VU blocks in UIC 918.3 tickets now though.