Base extractor filters on MIME types rather than our own type enum
This is mostly migration scaffolding for now, as neither the consumer
nor the provider side has been adjusted to use MIME types yet, but we
can adapt both sides independently now.

This is part of a larger set of changes phasing out the use of
ExtractorInput altogether, as that ties us to a hardcoded set of
supported types.