Pick a better modification time for pkpass content if we have one
The "relevant date" property isn't ideal for that, as it is unreliably
filled and even when set correctly refers to the time of the booked event
rather than the booking time. That's still useful for context, but for
modification time (which we need for proper update handling), the
surrounding document can be more useful.