Add JS API to retrieve child nodes matching a given filter
This will be the replacement for the few remaining extractor scripts still
relying on the unintended behavior of being iterated over all detected
barcodes even when that isn't asked for in their filter rules. In a few
cases we cannot change that (which would generally be better), as there
is nothing strongly identifying those barcodes (e.g. just containing
random ids).