Add support for compressed images outside of PDF files
This works for anything QImage supports, although only PNG is registered
so far, as that's practically the only format in use for barcodes.

This is mainly needed for porting KItinerary Workbench for now (which did
this handle this case manually so far), but we'll need that for supporting
images in HTML emails anyway.