Add basic bit layout for ERA SSB v2 tickets
This is (mostly) what Trenitalia is using, although in the second half
this isn't matching their content it seems. Even if that turns out to not
be a bug in our code this is already worth it, as this manages to properly
decode international destination codes, unlike our previous code.