Correctly decode HTML MIME nodes which specify codecs in HTML and MIME
We must not apply the same codec twice there, as the input to the HTML
encoding is UTF-8 in that case, not what might be specified in the HTML
content-type header.

This is a problem for e.g. Japanese and Korean codecs in particular, where
doing this wrongly completely destroys the content.