Use KI18nLocaleData when available
This duplicates a bit of code we can remove again once we depend
unconditionally on KF 5.88. Doing this now already however allows
for an easy side-by-side comparison of the results.

There are minor changes in places with ambiguous timezones, and
since country lookup is no longer tied to timezones, we now also
get proper results for Thailand and northern Vietnam.