Produce better results for just scanning a DB domestic UIC 918.3 code
This fell back to the generic UIC 918.3 ticket extractor so far, as we
assumed there always to be a PDF and thus didn't bother with generating
better results for just the barcode. Doing that was also a problem in
the past as those codes don't include the route details, but meanwhile
we have the triangular merge logic to deal with that also when getting
the full route details later.