Add workaround for trainline.eu emails containing wrong SNCF ticket tokens
This seems to happen in case there are multiple SNCF legs with different
ticket tokens. The JSON-LD data in the HTML part of the email then uses
only one of those for all legs.

By dropping ticket tokens from the HTML part we end up using those found
in the attached PDF tickets, which are from SNCF itself and thus correct.