Add initial ERA Flexible Content Barcode (FCB) structures
This is far from complete and merely covers the first few elements of
the full FCB specification. It's sufficient to validate the ASN.1 uPER
decoding code on real-world samples though.

This is still very verbose and repetitive, we can hopefully improve
that by macros or even code generation from the ASN.1 specification.
This is also not yet able to handle optional primitive type fields
properly.