Add basic uPER element macros
This reduces the boilerplate needed for Q_PROPERTY declarations, and it
adds proper handling of optional elements.

This can potentially expanded to automatically count optional elements
with a mechanism similar to the one used for generating the comparison
operator in the datatype classes, which would avoid the error-prone
manual indexing.

We probably want something similar for the implementation as well, that
is also still rather verbose.