Automatically determine bitmap indexes for optional uPER elements
This uses the same trick as the auto-generated comparison operators in
the JSON-LD datatypes, which in turn is inspired by Verdigris.

This further reduces error-prone boilerplate code, but it also means we
cannot skip not yet implemented elements anymore, that would now fail to
decode. This is something we must not do in the final code anyway though,
it only works for now when knowing those fields also don't occur in the
test data.