Make ZXing mandatory
We lose too many functionality without it, and the only reason it hasn't
been mandatory yet is that not all CI configuration had it available. That
has been fixed now though.

This only switches the CMake requirement, cleaning up the conditional
code, alternative test data, etc can be done in a few weeks if this
doesn't cause any unexpected fallout.