Start of a Protocol Buffers stream reader
This is intended for use with protobuf encoded data for which the .proto
file is largely unknown. We have that for example for SBB ticket QR codes.
For that we are only missing support for decoding strings and sub-messages
now, and probably some convenience methods to navigate through the fields.