Add method to check whether two reservations refer to the same incidence
That is both are for the same thing but for different people. That's
useful for grouping reservations together.

This code exists in three different places at least but neither of those
handle the hotel reservation special case correctly. So do this properly
here and make it available as public API.