Add support for running extractors on HTTP responses
We'll need that to support importing tickets via API calls to the
respective providers. The response data can contain nested elements
and ticket barcodes etc, so this is best processed with all the
infrastructure we already have here.

This is mainly intended to be run from QNetworkReply instances, but
there is also basic support for consuming HAR archives. The latter is
useful for development and testing.