Add ActivityPub event extractor
This is very similar to the schema.org Event structure, we just need a few
minor transformations.

In theory this would probably need the fully JSON-LD schema-aware
normalization stuff, but in practice this seems to be good enough
without that.

This will be used by Itinerary to import events from ActivityPub sources
such as Mobilizon.