Handle repeated train constraints in SBB ticket barcodes
That happens if the train is bound to a specific journey with multiple
legs. That's of limited use as such (we'd want this to split up into
individual entries), but it's all that is in the barcode itself. An
additional extractor script expanding that from the corresponding email
should be able to do that.