Add a workaround for fixing broken UIC 918.3 payloads in Renfe tickets
There are a few fields with slightly wrong content or off by one lengths
compared to what the specification demands. If we detect that, rewrite
the raw payload data accordingly so the regular parser works sufficiently
far to produce useful output.

There's more errors around content text encoding, but the existing
workarounds for Denmark seems to cover that for the samples I have.