Ignore ERA FCB UIC station codes in Germany
Several operators seem to use different code tables, and most of them
don't match what we have in Wikidata. That results in all kinds of broken
assumptions and behavior in Itinerary.

Until that is resolved, we only consider the country information from
those (which in case of Germany also implies the timezone), but not the
coordinate.