Make the datatype implementation macros externally usable
Itinerary uses the header part of this internally as well but since we
changed the comparison operator its implementation side of this is out
of sync. Avoid that problem entirely by using the smae code everywhere.

Besides installing a few more headers this also requires using fully
qualified type names as this is used in a different namespace externally.