add: Ticketportal event ticket extractor
This commit adds an extractor for Ticketportal (https://ticketportal.cz) so called hometicket .pdf ticket.

This extractor only supports the czech localization as that was the only one i had to test with. Ticketportal operates also in Slovakia, Hungary and Bulgaria.

Problem with this extractor and ticketportal itself is  that that they dont write the name of the event in text but as part of an .png. Another problem is that ticketportal allows for unstandardised venue location strings.