Use the generic subjectOf property for attaching Apple Wallet passes
We have our own extensions on Reservation sub-types for this so far, but
that is rather limiting when it comes to attaching passes to other types
of objects. Instead, let's use something more standard-compliant and
applicable to all types.

This allows Itinerary to support tickets and program membership cards
in pkpass format.