Make sure we will always create an extractor document node
To the very least any input data can always result in a single binary data
application/octet-stream node.

We failed to do that until now when the input is detected as a specific
but unsupported mimetype via the mimetype database autodetection, and went
unnoticed as in those cases we tend to have no extractor handling that data
anyway. It's a problem when consumers build additional extraction logic
on top of this though, such as Itinerary's health certificate detection
(DIVOC certificates were detected as ZIP files and thus resulted in an
empty document node tree).