fcmtravel: Extract arrival/departure terminal information if available
FCM reservations have this information, but of course it's optional
depending on if/when the reservation was booked. It's on the same line
as the other arrival/departure information which is a bit annoying, so I
made the regex more greedy and then added another regex match to extract
the actual terminal identifier.

Note that I only had an English FCM reservation, someone else can add
the Swedish translation later.