Added an extractor script for hostelling scotland reservations
Wrote an extractor script for hostelling scotland youth hostel
reservations. Data is extracted from a pdf attached to the reser-
vation e-mail sent by reservations@hostellingscotland.org.uk
I had two bookings available for 2 different youth hostels and
tested the script with both. Those are reservations for one
hostel each. Unfortunately, the reservations do not include
a house number in the street adress.

There seems the option that multiple hostels can
be booked in the same reservation (based on their website).
However, I could not implement that due to the lacking of
such a reservation.