Several fixes for ELB date/time parsing
Compute dates correctly in relation to the context date, ie. assume the
emission date is in the past but the travel times are in the future, so
that we at least end up with a correct order of all dates in case of
ambiguities.

Also, expose the validity dates as date/time values to QML, as QML messes
up pure date values when converting back and forth as it's internally
stored as a time value in the local timezone.

Also, extract ELB validity times for Eurostar.