Yet more fixes for the United Airlines parser
One of the issues is caused by the MileagePlus table in the e-mail
that interfered with the arrival/departure date parsing, because
the XPath query was too broad. And then when it came to traveler
name parsing, I'm not sure how it ever worked (?) but now it should
correctly fill out the Person object.

(cherry picked from commit 2a22c345615465a894c9bd6eda6b23dc063721c6)