Merge common parts of all elements of the same incidence
This is particularly relevant for tickets split between e.g. the trip
and a reservation (common e.g. with ÖBB), and both parts containing
a different subset of information. With this those are now all in sync.

This also reduces the spelling variations of stops you might observe in
Itinerary in such a case.