Fix offset computation in overflow airport name string table
This didn't account for the trailing null byte of the first table,
so all access to the overflow one were off by one and thus not only
resulting in wrong strings but also violating the order precondition
of std::lower_bound.

This becomes more prominent with more data moving to the overflow
table with the upcoming database update.